Disappeared or maximally retried nodes should be job failures
As discussed: https://github.com/bacalhau-project/bacalhau/pull/2518#issuecomment-1576911394

It is a job failure if the `any` strategy cannot schedule the job onto
enough nodes. This makes sense because we have not been able to fulfil
the user's Deal.

This commit modifies the `all` strategy to have the same behaviour. If
the job cannot be successfully run on "all" nodes, the job now fails.