Fix jobs targeting all nodes being unable to retry
We have only tested this code in the CLI (where it's hard to generate
jobs that should retry) and in unit tests where we using faked peer
IDs. It turns out that with real peer IDs we were not comparing them
correctly.

We now compare them correctly. This includes some end-to-end tests for
jobs targeting all nodes.