job info consumer
A separate application that consumes job info events and feed them into postgres database. It doesn't expose any interface to query or visualize the data as that is currently possible through grafana and looker.

This is a replacement to our current dashboard. It follows a very simple schema that should withstand changes in our job and status specs. Separate postgres views are created to provide more valuable summary of the data which can be updated with future changes in job specs