fix TestEnvWriter
the test case is veryfing WriteRunInfoFile, which creates readonly bacalhau.run file in each run, but doesn't delete it. The next test invocation will fail because the file already exist and not writable. The fix is delete the file when done, which is what `bacalhau serve` and `bacalhau devstack` do.