Show node ranking failures in the CLI
Previously when so many nodes fail ranking that not enough are
available to run the job, the CLI would just print an unhelpful
"not enough nodes to run job" message. This makes it hard to diagnose
why the job isn't suitable.

This commit adds a reason to the node rankings, and in the event of
total job failure displays the reasons for each node in the CLI.