Add an etcd-backed priority queue
Adds a priority queue that uses a prefix to store objects whose key is a
sequence. The prefix contains the priority of the item being stored and
so when we retrieve it (Get) we get the first lexicographic item - and
we claim it by deleting it from the store.

Mostly based on a recipe provided by etcd.