Fix container entry points (#2302)
Fix behaviour of cmd and entrypoint for bacalhau docker run

The problem:
"From a user perspective, if you define either an entrypoint or a command in a container, the defaults are overwritten with an empty value if you pass no arguments to bacalhau. This is unexpected, from a user perspective. Default container entrypoints and commands should be used."

Using the entrypoint flag will now correctly override the docker image's default entrypoint. Note that the dockerfile's CMD will be ignored if the entrypoint flag is used. Any arguments appended to the CLI will be interpreted as the new CMD.

Similarly, if the entrypoint flag is NOT used (which will probably be most cases) any arguments appended to the image name will replace the image's default cmd. 

If no arguments are provided, and the entrypoint is not overridden, bacalhau docker run will use the image's entrypoint and cmd, as expected.

tldr: bacalhau docker run works like docker run