implement evaluation worker (#2616)
Worker is a long-running process that dequeues evaluations, invokes the
scheduler to process the evaluation, and acknowledges or rejects the
evaluation back to the broker. The worker is single-threaded and
processes one evaluation at a time. An orchestrator can have multiple
workers working in parallel.

Future PRs will include actual implementation of schedulers, and linking
things together