Implement GetExecutionCount
Implements GetExecutionCount by maaintaining an in-memory counter of how
many executions are in a particular state. As this is inmemory it is
filled at runtime when the compute node creates its execution store.
Retrieving execution counts will block until this startup process has
completed.

The execution store interface was improved during this work to require
that GetExecutionCount receive an ExecutionState so that it is possible
to ask for a count of any particular state and not just
ExecutionStateCompleted.