Compute Object Store (#2597)
Currently execution data is held in memory by the individual compute
nodes. This means that should a compute node be restarted, all knowledge
of the work it was doing is lost. As we move toward different types of
jobs, where we expect some to always be running we need to persist this
state so that compute nodes can resume work from where they left off.

To provide this resilience, this PR introduces a local persistent store
for compute nodes using boltdb key-value database (using the etcd
variant). This new execution store is interchangeable with the current
inmemory store, implementing the same interface and passing the same
tests. The inmemory store is currently the default but the kvstore can
be enabled using environment variables:

```shell
# Enable the boltdb storage, alternative option is inmemory (or "" for the default)
export BACALHAU_COMPUTE_STORE_TYPE=boltdb
# Specifies the file to be used by the compute node's database
export BACALHAU_COMPUTE_STORE_PATH=/tmp/database.db
```

The database created is a single file, stored in
`BACALHAU_COMPUTE_STORE_PATH` or the config directory if this is not
specified. When not specified, the filename structure is
`NODEID-compute.db` and is persistent. Any databases created during
tests are, and should be, cleaned up during teardown.

As part of this work the execution store model has changed to allow
`GetExecutionCount` to receive a `store.ExecutionState` for which the
count is required. For this store it is held in-memory and loaded at
runtime from the current state in the database at launch time. This
background task might provide the initial basis for start-up compaction.