Ensure tools required for go generate are installed
When running `make generate` to run `go generate` over the source code,
some of the required tools (mockgen and stringer) may be missing.

Adds a new target, upon which `make generate` depends to install the
tools if they are not currently available (as determined by `which`).

Closes #2651

Please enter the commit message for your changes. Lines starting