speed up node discovery during startup (#2653)
Node discovery during startup takes long (~30 seconds) for the nodes to
properly connect to libp2p gossipSub and advertise itself to the
network. This is slowing down our tests significantly. Even though we
were allowing the node to eagerly advertise itself as soon as it starts
up, it wasn't enough for our tests as the gossipsub network is not well
established by that time and we don't have a clear way to know when the
gossip peers are well connected.

This PR introduces a configuration to allow nodes to _continuously_
eagerly advertise themselves during startup with shorter
`EagerPublishInterval` up to `EagerPublishDuration`, after which it will
switch to standard intervals.

These are the current configurations:
```
var DefaultNodeInfoPublishConfig = routing.NodeInfoPublisherIntervalConfig{
	Interval:             30 * time.Second,
	EagerPublishInterval: 5 * time.Second,
	EagerPublishDuration: 30 * time.Second,
}

// speed up node announcements for tests
var TestNodeInfoPublishConfig = routing.NodeInfoPublisherIntervalConfig{
	Interval:             30 * time.Second,
	EagerPublishInterval: 10 * time.Millisecond,
	EagerPublishDuration: 5 * time.Second,
}
```

This will allow nodes to be discovered in milliseconds during tests
instead of 30 second per test scenario!

This change is also allowing us to remove `IdentityNodeDiscoverer` which
was only needed because of early node discovery in some tests, but did
not provide the full node information required by mosts tests, such as
supported engines and label selectors