Preserve storage path and name when copying
Previously, our copy algorithm did not preserve the name or path of a
passed StorageSpec, meaning that the mount point was then no longer
available when a spec was copied between drivers.

Our copy algorithm now preserves these. It does not preserve the
metadata because this is defined as "driver-specific" and hence it is
up to the destination storage driver to supply it.

Resolves #2666.