Transport errors to the client in main message struct
Currently messages are sent via the control channel, and this runs up
against an unclearly documented limit of 125 bytes for the control
channel message.  This means that all errors end up as abnormal errors
at the client end, something we should never send ourselves.

Now errors are included in the message structure, if present it is
printed to STDOUT and the process is cancelled.

No changes to access policy have been made yet.