refactor: executor interface and storage dependency (#2675)
### What
The goal of this PR is to simplify the implementation #2637 by removing
the requirement of configuring storage providers for executors. It
includes several refactors to the bid strategies, executors interface,
and corresponding implementation.
- Executors no longer contain storage providers. Instead the compute
node responsible for orchestrating executors maintains the set of
available storage providers. The compute node makes storage available to
executors as local volumes.
- ParallelPrepareStorage now returns a slice of prepared storage
(instead of a map). This allows the returned `PreparedStorage` type to
be serialized and passes across RPC boundaries.
- `NewModuleLoader` will not fetch remote wasm modules, and instead
operates over a slice of `PreparedStorage`.
- This change was required to remove the Storage Provider type from the
wasm executor which previously fetched remote content from `Job.Input`
during execution.
- Modifies the `Executor` interface by removing methods related to
providing storage and reducing the complexity of parameters passes to
the `Run` method.