introduce scheduler plan executors + migrate requester (#2681)
Add implementations for `orchestrator.Planner` that executes the plan
proposed by the `orchestrator.Scheduler`. To make sure things are
working as expected, This PR also migrates the `requester` to use the
new path of scheduling jobs through the `orchestrator` package, and now
the `requester` is mainly an endpoint to handle job submissions and
compute callbacks.

Planner implementations include:
- StateUpdater: This planner updates the job store with the desired
changes resulting from the evaluation.
- ComputeForwarder: This planner notifies compute nodes with the
proposed changes and updates the job store with the final observed state
if the notification is successful. It handles communication with the
compute nodes.

In addition to that, `Execution` is now represented by two states. A
reminder that an `Execution` represents a mapping of a job to a compute
node, where a job can have multiple active executions based on the
desired executions count. The two states representing an Execution are:
- `State`: Which is the existing state and represents the latest
observed state of the execution
- `DesiredState`: Represents the desired state of the execution as
provided by the scheduler.

The following sequence diagram shows the flow of an execution from the
time it is created until it is completed, along with the different
components that can mutate its state:

![hPBBhjiW48RtF0KwArlgUG5KLN5waTf55eswouR0jAN1W33aylOTNDl4JfqSIzaam7mpyy_yNZM1-n1hJHp-Aw4SDu6-V3RIMMN22-KDXUuoGOvSHnwKDJkpzlO0_X-CpsFmczrlW_NGXfjq6iUtuZV8gBlyfhzDeC03xCmF9tdeoVJKFOpXZxPsCU5Vh3znBxi4v-tvGWXv_J2CnEWsxciLK2FX8gWZ5ilpzWUA6QrmHeK7](https://github.com/bacalhau-project/bacalhau/assets/7647060/15f746e3-f44e-4a55-97eb-6a4d3c99372a)

At a high level, the scheduler is responsible for updating the
execution's desired state, while other components are responsible for
updating the observed state as follows:
- The endpoint receives requests from end users and callbacks from the
compute nodes. Its primary responsibility is to update the execution's
observed state and notify the scheduler of the new state by enqueuing an
evaluation.
- The scheduler polls an evaluation and creates new executions or
updates the desired state of existing ones. It does not modify the
observed state. It does that through the help of the state updater.
- The compute proxy receives the plan and updates the execution's
observed state after notifying the compute node of the change.