Promotes EnvironmentVariables into the spec.
Currently the environment variables are held (differently) within the
engine specs, but they're really a property of the environment where the
executor will be run - and so it feels like it might make more sense for
them to be part of the JobSpec. They're also held there as a
map[string]string rather than []string.