introduce ops job type, a batch job that targets all matching nodes (#2690)
Other key changes with this new job type:
- Requester can now ask a compute node to run a job immediately when
there is available capacity without having to first ask the requester
for approval (ie BidAccepted or BidRejected). `OpsJob` uses this pattern
as there is no reason to approve or reject these types of jobs.
`BatchJob` still follows the existing approval approach. The code is
still ugly because we are responding to bid requests async, which I am
planning to revisit when Simon is back. Also will revisit changing the
method names
- The requester doesn't retry when a node fails an `OpsJob` as there is
no value in attempting to retry on another node in the network since
we've selected all matching nodes