Persistent Jobstore (#2650)
Currently we have only an in-memory implementation of a jobstore, and
this isn't very resistant to failures. This PR introduces a boltdb
backed jobstore to ensure jobs and their state and remember across
restarts of the requester node.

The jobs in the database are structured as follows, with uppercase names
reflecting dynamic values, and `key:val` showing key-value mappings. In
cases where the the path does not end in a `key:val` entry, these are
the index buckets described below.

### Buckets

```
Jobs 
    |---> bucket <JobID>
                |----> key:Spec value:{model.Job}
                |----> key:State value:{model.JobState}
                |----> bucket job_history
                                       |----> key:seq value:{model.JobHistory}
                |----> bucket execution_history
                                       |----> key:seq value:{model.JobHistory}
                |----> bucket executions
                                       |----> key:execID value:{model.ExecutionState}
                |----> bucket evaluations
                                       |----> key:evalID value:{models.Evaluation}
```                  


### Indexes

Some of the entries in the database, such as the list of Job IDs for a
given client-id are actually indexes. These are structured as buckets,
with the root being "idx_NAME". So for client ids we have "idx_clients"
and then the next bucket showing the "CLIENT-ID" and then a list of
buckets showing the jobs from the client-id. They could have been
structured as buckets labelled "clientid;jobid" and found with partial
matches but this requires we iterate over all of the buckets. This way
we can do a direct lookup on the clientid, resulting in iterating many
fewer buckets.

Current indexes are 

* TagsIndex =  tag -> Job id
* ProgressIndex = job-id -> {}
* ClientsIndex   =  client-id -> Job id
* EvaluationsIndex = evaluation-id -> Job id


### Bucket paths

This PR also introduces bucket paths which are used for finding buckets
from strings like "root.bucket.with.subbucket". This performs the same
number of lookups as doing this work each time, but also provides the
option to create any buckets in the path that don't yet exist.

### Watching events 

The jobstore now contains a method (Watch) which will provide the caller
with an event channel filtered to their requirements. By specifying the
event types (Create, Update, Delete) and the target objects (JobWatcher
or ExecutionWatcher) the returned channel will receive events that only
match those criteria. The `WatchEvent` received on the channel also
contains a timestamp, and a `[]byte` format json-encoded of the related
object type.

|Watcher type|Event type|Returned object|
|--|--|--|
| JobWatcher | Create | Job |
| JobWatcher | Update | JobState |
| JobWatcher | Delete | Job |
| ExecutionWatcher | Create | Execution |
| ExecutionWatcher | Update | Execution |
| ExecutionWatcher | Delete | - |
|EvaluationWatcher|Create| Evaluation |
|EvaluationWatcher|Update| - |
|EvaluationWatcher|Delete|Evaluation|


Resolves #2645

---------

Co-authored-by: Walid Baruni <wdbaruni@gmail.com>