Allow maximum job execution time to be configured
Previously Bacalhau had some hard-coded values for maximum job time.
Having a maximum timeout makes sense on a public network where
non-paying users should not be able to hog capacity. For private
clusters, the behaviour should be configurable – users may wish to
run jobs that span many days.

In this commit we provide a new config flag `--max-timeout` respected
by `bacalhau serve` that allows these maximum job times to be set by
a node operator. It also changes the default when the flag is unused
to jobs having no maximum execution time.

The behaviour of the Bacalhau demo network is unchanged.

The timeout value has been changed from a floating point number of
seconds to an integer number of seconds. There are no instances of
anyone submitting jobs with fractional seconds on the public network.