Remove coupling of host and api env vars
Currently users must specify both the `BACALHAU_API_HOST` and
`BACALHAU_API_PORT` settings and both are used to build the URL of the
requester node to use.

This PR removes that coupling and allows for them to be specified
individually, where the default is still used if that component is not
specified. e.g. you can change `BACALHAU_API_HOST` and it will continue
to connect on the default port (1234). Or you can specify
`BACALHAU_API_PORT` and it will connect on the default host (the public
network).  Specifying both continues with current behaviour.