Allow the execution store to quickly find 'live' executions
When executions are in an executing state, then the execution id is
added to a live-executions index. When it exits an execution state, it
is removed from the index.

This is implemented to allow the compute node to try and reconstitute
the state of the node when the requester last ran.