Adds a container for the startup tasks in compute node
When the compute node starts, we want to run some tasks which might be
cleanup tasks, or as in this case, resetting the state of the node.  As
we don't know the reason we are starting, we check the execution store
for any live executions and determine what we should do with them.

For short-lived jobs we want to mark them as failed, but currently we're
only able to cancel them as the fail state is closely tied to the
synchronous Run() method.  For longer running jobs, we currently cancel
the job and then attempt to run it as we have no way of reconnecting to
the underlying process, although pluggable executors should enable this.