Adds support to the requester public API
Adds a `--tlscert` and `--tlskey` setting for the server certificate and
key (respectively). When testing with self-signed certificates, it will
be necessary to ship the CA certificate to any clients wishing to
connect to the API.  In situations where hostnames resolve and a CA is
available to resolve them (e.g. production) then the client will not
require the CA certificate.