Add TLS support to client
When using the client to connect to a https bacalhau API there may be
cases where the TLS certificate is not trusted because it was
self-generated in a private network.  To enable the client to connect to
this server without complaint, there are two new global command line
options.

`--insecure` will allow the client to connect to a HTTPS bacalhau
without verifying the certificate of the server.

`--cacert [path to file]` allows the client to use the provided CA
certificate to verify the certificate is valid.