New model types (#2719)
Introduce new model types and job specs that unblock more job types,
extensible engines and storage sources, and updatable jobs.

### Models Summary
#### Job 

`Job` is an internal and common representation for the different job
types submitted by end users. Users won't submit `Job` directly, but
instead will submit different request and spec depending on the job type
(not introduced yet), and then those requests will be translated to a
`Job` spec, which allows introducing new job types with no or minimal
change to existing schedulers or executors. Changes to existing `Job`
include:

- `Name` field defined by the user, which will allow updating existing
jobs if a job with the same name already exist for the same user
- `Spec` has been replaced by a list of `Tasks` to enable running
parallel or sequential local tasks per job in the same compute node.
(e.g. parallel task to publish results as they are made available by the
main task, or a task at the end to publish results which can replace our
current publishers). Won't introduce this capability with this PR.
- `JobState` type has been replaced by a single `State` field in the
`Job` type itself.

#### Task

Represents the previous `Spec` type, which the following changes:
- `Engine`, `Publisher`, `Inputs` and `Outputs` are all represented by a
the same flexible `SpecConfig` type, which has a `Type` string field,
and a `Params` map of strings to interfaces. This works nicely with
existing `Publishers`, new pluggable executors, and will unlock the same
flexibility for storage specs
- Previously `Engine`, `Publisher` and `StorageSource` names were
_indirectly_ case insensitive as we ignored the case when deserializing
those names into their enums. As I've switched to strings instead of
enums in the new models, extra care was required to maintain this
behaviour. Switching to snake case for future names might simplify
future work (e.g. `urlDownload` -> `url_download`)


### Main Changes
This turned out to be a very large PR, but most of the changes are
because of using the new types and fixing tests. These are the main
changes to help with the review:
1. `ExcludedTag` and `IncludedTag` are now represented as strings
instead of a dedicated type. Also `inmemory` job store now gives higher
precedence to `ExcludedTag` so that if a tag shows in both `IncludedTag`
and `ExcludedTag`, it will be excluded. This is a similar behaviour to
boltdb implementation, and can change results of `bacalhau list` for
some users
2. `JobSelectionPolicy` struct has been moved from `model` to `node`
package as it represents configuration for _some_ of the bidding
strategies. We might remove this struct with proper configuration setup
3. Moved marshallers and providers from `model` package to their
packages under `lib`. Providers now accept a string a key, and does case
insensitive matching as explained above
4. Compute `store.ExecutionState`, which is almost a duplicate of the
job store's `models.ExecutionState`, has been merged and replaced with
`LocalexecutionState` that wraps `models.ExecutionState` along with
additional fields to track the local compute's state