Use priority field for compute node execute buffer (#2765)
Currently the executor_buffer which maintains the queue for the compute
node running tasks where there are available resources to do so. This is
currently a FIFO queue where jobs are processed in order, although it is
able to skip entries in case a later job has fewer resource requirements
which can be satisfied now.

This PR changes the executor_buffer over to using a priority queue so
that jobs are processed in order of priority (currently the job priority
field). If dequeued tasks have resource requirements which can't
currently be met, they will be requeued once the iteration has
completed.