Restart long running jobs after compute node visibility change (#2734)
Compute node visibility will change over time. This may be caused by
network partitions, intentional lack of connectivity, or short term
compute node failures (whether h/w or s/w failure).

When visibility is restored, there is a risk that tasks that should be
running will be lost. There are several options to avoid this happening.

1. Abandon all existing state, launch a fresh instance of the compute
and inform the requester node that the previous node-id is no more. The
requester node will send work.
2. Similar to 1, but relaunch the same instance, and just inform that
requester node that all tasks that were running have failed.
3. Have the compute node determine what to do with tasks, cancelling (as
we can't yet record an async failure) any short lived tasks, and
cancel/restarting any long-lived tasks.

This PR implements option 3, but this may not be optimal because:

* The requester node is informed of state changes via the compute node
Cancel/Run calls, and these may not happen for a variety of reasons
* The state changes may happen, but be out of sync with the requester
node resulting in needing to take more action locally
* We don't know the state of the underlying execution environment, which
may result in orphan processes eventually consuming node resources


This PR adds a `Startup` structure to contain tasks that we want to run
on the compute node startup (e.g. see other tickets about cleaning up
various job inputs). Here we have added a check on all of the executions
that the execution store believes are running, and then based on the job
type, we decide on what action to take.

For short-lived jobs we ideally want to mark them as failed, but
unfortunately that is only available as part of the current Run() flow -
so instead we use `Cancel()` with the intention to change it later.

For long-lived jobs we want to be able to check and see if it is
running, after which there is nothing to do, or whether we need to
re-Run the execution. Unfortunately we don't have a mechanism yet to
tell what the state of the underlying process is (e.g. docker, wasm etc)
and so we instead `Cancel()` and `Run()`.

The solution for neither job is ideal, but it may be possible to extend
the pluggable executors to support the desired behaviour.