Use a temporary folder for the devstack repo
Whenever we create a devstack, the result is lots of files stored in the
repo for each compute and requester node started that never get cleaned
up. This PR changes the devstack to use a temporary directory for its
repository.  This means we can auto cleanup the entire repository when
the devstack is torn down.

TODO:

It would still be useful to force a specific location for the repo but
without the burden of a default setting. e.g. if --repo is specified, we
should use it, but we don't want the default value at all.