Default to persistent stores (#2779)
Now prod, test, dev, staging envs will use boltdb by default for
persisting state, with devstack continuing to use inmemory. This is easily changed for
devstack using the `COMPUTE_STORE_TYPE`, and `JOB_STORE_TYPE` environment
variables set to "boltdb".

In addition to making them default, it also changes the local of where
the database files are stored. They continue to be part of the config repo, but are
stored according to their node id and type. For each database file
(executions.db and jobs.db) they can be found in a folder similar to
`~/.bacalhau/{node-id}-{node-type}`.