use go-chi api router + refactor api package structure (#2772)
### Summary
This PR refactors and re-organizes our existing API packages as follows:
1. Use `[go-chi](https://github.com/go-chi/chi)` for routing requests,
middleware orchestration, and json rendering of responses
2. Move API endpoint implementations from `requester` and `compute`
packages to under `publicapi/endpoints/` to have a single place for all
our http endpoints
3. Have a unified client implementation that supports all endpoints.
This also introduced `publicapi/apimodels` package as a single place for
all our request and response types
4. Introduced new set of middlewares and changed how we log requests,
migrate legacy paths, recover from panics, and assigning unique request
IDs
5. Improved APIConfig and now using functional options

### Identified Issues
Almost all of our APIs are using `POST`, where when `GET` is more
suitable, such as `list` and `states`. The new router made this more
obvious. Will address this in the new set of APIs

### Next Steps
Next step is evaluating `echo` framework to implement the handlers as
discussed with Forrest. Most of the work in this PR is re-organizing our
packages, and go-chi was easy to use since it is compatible with
standard handlers that existing APIs uses

ENG-54