use echo web framework (#2775)
This PR migrates from chi router to echo framework. 

The main benefit that I see with echo is the central error handling
where the http handler signature `func(c echo.Context) error` returns an
error instead of the standard `func(w http.ResponseWriter, r
*http.Request)`. This allows injecting an error handler that captures
all returned and unhandled errors, process and transform them before
returning to the user. I am not utilizing the error handler in this PR
because of how we are handling errors in the existing code. A similar
behaviour is still possible in chi by passing an error handler and
calling it inside the http handlers. We are already doing something
similar, but it can be improved
https://github.com/bacalhau-project/bacalhau/blob/chi/pkg/publicapi/endpoint/requester/endpoints_cancel.go#L37

The drawback is deviating from the standard library http handlers, which
will make migrating away from echo to be more challenging, and that we
will have to wrap 3rd party middleware that were implemented for the
standard library and try to make them compatible with echo.

In my opinion, I don't see strong value with using echo and deviating
away from the standard library handlers. I see go-chi more than enough
for our needs, but I don't have strong opinion about this and I am okay
with the switch given Forrest's positive experience with echo in a
previous project.