Add option to specify location of devstack config
By default, the devstack config is now stored in a temporary folder,
which is removed when the devstack is shut down.

There are cases, however, where it is useful for the devstack config to
persist beyond the live of the devstack, allowing for inspection of the
values in the repo, or access to the databases.

By specifying `--stack-repo PATH` it is now possible for the devstack to
create a config repo in a specific location that is not removed when the
devstack closes.