Add worker pool to process new evaluation messages
To be able to process new evaluations being created, we need to be able
to do so without blocking the process generating the evaluations (and
notifications).  This commit adds a worker pool which allows us to
process messages of a known generic type, and control how buffered (or
not) the input channel is, and how many worker routines will exist to
process the messages.  We must be aware though that if we are writing to
the database, there will be a lock on the single db writer.