Async Compute Executors  (#2778)
### What
This PR modifies the executor interface to include the following
additions:
```go
Start(ctx context.Context, request *RunCommandRequest) error
Wait(ctx context.Context, executionID string) (<-chan *models.RunCommandResult, error)
```
It then updates the Noop, Wasm and Docker executors to satisfy the new
interface requirements. Each of these executor changes needs some review
as the changes are more than pure mechanical changes. A `handler`
pattern has been implemented to manage each executors active executions.

### Call outs
- Previously when a docker container was passed an invalid
entrypoint/command it would return an error rather than start the
execution. There were some tests that made assertions on this. Since
starting the docker container is the same as running the job, and we
want to run jobs async, it is no longer the case that these errors are
returned immediately. They are instead shown when describing the failed
executions state. I have modified the required tests to account for this
change in behavior.
- The ExecutorBuffer triggers job timeouts via context cancellation by
passing the Run method a context with a timeout. I am dubious this will
continue to work as intended when operating against pluggable executors
and suspect it will instead end the client connection to the plugin
server. For the time being I have worked around this
[here](https://github.com/bacalhau-project/bacalhau/blob/9834a589ee491f74f65b5e3b603a904348ab8453/pkg/compute/executor.go#L288)
and added a TODO comment. IMO we should instead explicitly call the
`Cancel` method on an executor to cancel its operation instead of
relying on the passed context. As mentioned in the TODO comment there is
more context on canceled job states
[here](https://github.com/bacalhau-project/bacalhau/pull/2705#discussion_r1284228115)
- As Ross has already brought attention to
[here](https://github.com/bacalhau-project/bacalhau/pull/2778/files#r1314919569)
`PrepareRunArguments` now returns a function that may be called to
trigger cleanup of input volumes. We should continue discussion in the
linked thread regarding further changes to this design. I believe this
is related to some work @simonwo is landing regarding restarts.
- WriteJobResults function signature has been modified to no longer
return an error, instead the error is embedded (as it already was) in
the returned `RunCommandResult` structure. This change was made to
simplify logic around checking errors returned by the method.