Adds manual TLS certificate support to requester (#2800)
Adds manual (as opposed to automatic ACME) TLS certificate support to
the requester node.

By default, if none of the following options are used, the server will
continue to serve its API endpoints over HTTP. We should default to
HTTPS with a special opt-out, but we want to support some level of
backwards compability. For the next major version, this should setting
should be reversed and HSTS implemented.

In cases where ACME/autocert is not suitable, or not desirable, it is
possible to obtain certificates using the traditional route - create and
submit a CSR (Certificate Signing Request) to a Certificate Authority.
Or, if so desired a self-signed certificate can be created.

The server certificate and key are required, and the filenames passed to
the `--tlscert` and `--tlskey` parameters respectively.