fix failing to start when HOME is ureachable (#2805)
`bacalhau serve` and other CLI commands uses `$HOME` as the default repo
directory. Instead of finding `$HOME` IF the `--repo` flag or
`BACALHAU_PATH` env is missing, it tries to find `$HOME` preemptively
and set it as a default value to the cobra flag, and panics if it
doesn't find it.

This means all commands will fail if `$HOME` is unreachable even if the
user tries to set the repo manually. This PR fixes that