Buy more time with x86 osx builds
CircleCI are deprecating their x86 machines, on the basis that we can
then build on arm64 - although there doesn't seem to be a good answer to
"how do we test on osx x86?".

Some deprecations are happening in 2 weeks, which this PR hopes to
avoid, and will then allow us to continue until the end of Jan 24 at
which point x86 will no longer be available.

The change is actually just

```
  darwin-amd64:
    # Note - not an actual ARM macOS machine (https://circleci.canny.io/cloud-feature-requests/p/support-new-m1-arm-based-macs)
    macos:
      xcode: 15.1
    resource_class: macos.x86.medium.gen2
```

the rest is just formatter